Add comment rather than throw exception for airfield not in database
The nightflight module has an extensive database of airfield locations, but
it is reasonable to expect that some airfields will be tried that are not included.
These will include closed airfields and airfields without codes. If a night
calculation is attempted with such an airfield, it is better to add a comment then
throw away the calculation entirely.